requirements_gold.txt
# ===========================================
# GOLD TIER DEPENDENCIES
# ===========================================

# Core Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0

# Google API (Gmail)
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0

# LinkedIn API
linkedin-api-client>=0.1.0

# WhatsApp Business API
# Uses requests library for REST API calls

# Scheduling
schedule>=1.2.0
APScheduler>=3.10.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Logging & Monitoring
colorlog>=6.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
